rename internal to common
