+++ plugins/golang-filter/mcp-server/common/redis.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"context"
