+++ plugins/golang-filter/mcp-server/common/registry.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 var GlobalRegistry = NewServerRegistry()
