+++ plugins/golang-filter/mcp-server/common/server.go
@@ -1,5 +1,4 @@
-// Package server provides MCP (Model Control Protocol) server implementations.
-package internal
+package common
 
 import (
 	"context"
