+++ plugins/golang-filter/mcp-server/common/sse.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"encoding/json"
