+++ plugins/golang-filter/mcp-server/config.go
@@ -6,8 +6,8 @@ import (
 	xds "github.com/cncf/xds/go/xds/type/v3"
 	"google.golang.org/protobuf/types/known/anypb"
 
-	"github.com/alibaba/higress/plugins/golang-filter/mcp-server/internal"
-	_ "github.com/alibaba/higress/plugins/golang-filter/mcp-server/servers/gorm" // 导入gorm包以执行其init函数
+	"github.com/alibaba/higress/plugins/golang-filter/mcp-server/common"
+	_ "github.com/alibaba/higress/plugins/golang-filter/mcp-server/servers/gorm"
 	"github.com/envoyproxy/envoy/contrib/golang/common/go/api"
 	envoyHttp "github.com/envoyproxy/envoy/contrib/golang/filters/http/source/go/pkg/http"
 )
@@ -20,9 +20,9 @@ func init() {
 
 type config struct {
 	ssePathSuffix string
-	redisClient   *internal.RedisClient
+	redisClient   *common.RedisClient
 	stopChan      chan struct{}
-	servers       []*internal.SSEServer
+	servers       []*common.SSEServer
 }
 
 type parser struct {
@@ -44,12 +44,12 @@ func (p *parser) Parse(any *anypb.Any, callbacks api.ConfigCallbackHandler) (int
 		return nil, fmt.Errorf("redis config is not set")
 	}
 
-	redisConfig, err := internal.ParseRedisConfig(redisConfigMap)
+	redisConfig, err := common.ParseRedisConfig(redisConfigMap)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse redis config: %w", err)
 	}
 
-	redisClient, err := internal.NewRedisClient(redisConfig, conf.stopChan)
+	redisClient, err := common.NewRedisClient(redisConfig, conf.stopChan)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize RedisClient: %w", err)
 	}
@@ -80,7 +80,7 @@ func (p *parser) Parse(any *anypb.Any, callbacks api.ConfigCallbackHandler) (int
 		if !ok {
 			return nil, fmt.Errorf("server %s path is not set", serverType)
 		}
-		server := internal.GlobalRegistry.GetServer(serverType)
+		server := common.GlobalRegistry.GetServer(serverType)
 
 		if server == nil {
 			return nil, fmt.Errorf("server %s is not registered", serverType)
@@ -90,10 +90,10 @@ func (p *parser) Parse(any *anypb.Any, callbacks api.ConfigCallbackHandler) (int
 		if err != nil {
 			return nil, fmt.Errorf("failed to initialize DBServer: %w", err)
 		}
-		conf.servers = append(conf.servers, internal.NewSSEServer(serverInstance,
-			internal.WithRedisClient(redisClient),
-			internal.WithSSEEndpoint(fmt.Sprintf("%s%s", serverPath, ssePathSuffix)),
-			internal.WithMessageEndpoint(serverPath)))
+		conf.servers = append(conf.servers, common.NewSSEServer(serverInstance,
+			common.WithRedisClient(redisClient),
+			common.WithSSEEndpoint(fmt.Sprintf("%s%s", serverPath, ssePathSuffix)),
+			common.WithMessageEndpoint(serverPath)))
 		api.LogInfo(fmt.Sprintf("Registered MCP Server: %s", serverType))
 	}
 	return conf, nil
